gill_done/requirements.txt
absl-py==1.4.0
accelerate==0.20.3
attrs==22.2.0
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
diffusers==0.17.1
einops==0.4.1
exceptiongroup==1.1.0
filelock==3.9.0
fonttools==4.38.0
fsspec==2023.6.0
google-auth==2.20.0
google-auth-oauthlib==0.4.6
grpcio==1.54.2
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.6.3
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pluggy==1.0.0
protobuf==4.23.2
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
six==1.16.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tokenizers==0.12.1
tomli==2.0.1
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.11.4
torchvision==0.12.0
tqdm==4.64.1
transformers==4.30.2
typing_extensions==4.4.0
urllib3==1.26.14
git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git
Werkzeug==2.3.6
zipp==3.15.0
